refactor : app.js를 별도로 만들어 최종 component 관리
